requirements.txt
streamlit
pandas
tensorflow-cpu
opencv-python-headless
numpy